feat:文件服务重构

+++ services/server/file-server/src/main/java/io/github/architers/server/file/domain/entity/FileTemplate.java
@@ -43,7 +43,7 @@ public class FileTemplate {
     /**
      * 目录ID
      */
-    private Integer catalogId;
+    private Long catalogId;
 
     /**
      * 模板url
